requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-allauth==0.51.0
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
executing==0.10.0
filelock==3.8.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
identify==2.5.5
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
nodeenv==1.7.0
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
protobuf==4.21.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
traitlets==5.3.0
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
